clamp RGB colour after Oklab -> RGB conversion
lerping colours in Oklab can result in them going out of the (s)RGB range, thus producing negative numbers.
the rest of the code doesn't really handle this, so we have to clamp them after lerping before doing anything else with them.

in particular, the part that was getting glitched out was the communication code;
I don't really know the _exact_ reason, but I believe it's that the float was being truncated to an int32, and then wrapped around into a uint32.